Fix ITN027ST plots after Coulter CXP fix: were always using 0 as the minValue for simple log axis.  Use translated min value instead.
Add manually verified PlotTests junit test.  The data is forthcoming.

Add SliceLayout utility for cropping out individual plots from a large FLowJo generated layout image.

+++ flow/enginesrc/org/labkey/flow/analysis/model/DataFrame.java
@@ -110,7 +110,7 @@ public class DataFrame
             }
             if (paramInfo != null && paramInfo.getMinValue() != null)
             {
-                fn = ScalingFunction.makeFunction(fn, paramInfo.getMinValue().intValue());
+                fn = ScalingFunction.makeFunction(fn, paramInfo.getMinValue());
             }
             NumberArray from = data[p];
